feat: Add lazy loading for images and messaging feature
Refactor SessionCardView component to add lazy loading for images using the "react-lazy-load" library. This improves page load performance by only loading images when they are in the viewport.

Also, add a messaging feature to the SessionCardView component. This feature allows students to send messages to counselors. It includes a "FaComments" icon for messaging and a link to the "/student/message" page.